Add instancable shader wrapper
Magnum shader's represent the actual GL shader program and thus are not
meant to be instanced themselves. To homogenize the rendering process,
`ShaderInstance` wraps a reference to the shader program resource and
provides a layer that can be instanced and stores its own unique
uniforms. The uniforms are then updated via a CRTP function call before
drawing the specified object with the shader program.