Implement physical RCS
Added center-of-mass-aware thrust in MachineRocket. SysNewton got new
function to recursively compute the center of mass of a tree of
entities, as well as a function `try_get_or_find_rigidbody_parent()`
which was previously embedded in SysMachineRocket (it was needed in
SysMachineRCSController, so I moved it up to prevent code duplication).

AssetImporter now saves the `massdry` extra parameter as a member of
PrototypePart, which is then used to create an instance of ACompMass in
`part_instantiate()` so that it can be later used to calculate the
center of mass.