Add procedural inertia calculation, config reading
Added a bunch of functions to SysNewton to facilitate the runtime
computation of moments of inertia. Approximates mass distribution of
parts by computing their collision volume and assuming uniform density
throughout the volume.

Also added a temporary solution for config reading: machine
configuration is read into a `std::map<std::string, std::variant<int,
double, std::string>>` stored in the PrototypePart, which is passed as a
parameter to the machine upon instantiation. Will need to change later
if nested configs or array elements are added, but provides an
acceptable temporary solution for now.