Add per-object machines
The ability for sub-part objects to own their own machines has a number
of benefits and will facilitate future features like integral RCS
thrusters, or realistic resource mass distribution. Upon import, parts
are now recursively scanned for machines and PrototypeMachines are
stored in PrototypeObjects instead of PrototypePart. BlueprintObjects
also store BlueprintMachines, which are used for unique configuration
options, e.g. specifying the type or quantity of fuel in a tank,
choosing whether a piece of equipment is retracted or extended
initially, etc.