Revert ownership of protomachines to PrototypePart
In figuring out how to achieve per-object machines, the ownership of
PrototypeMachine information was moved to PrototypeObjects; this commit
reverts that change so that the PrototypePart once again stores the
machines, and the PrototypeObjects now just store indices to this array.

SysVehicle still tracks the association between machine and object, but
only saves the indices from the PrototypeObject and uses it to index the
PrototypePart's machine list.