Fix crash at exit
Crashing was caused by resources being freed before the DependRes
references that pointed to them. However, the resources must be freed no
later than just before the OSPMagnum object is freed, because resources
which depend on the GL context must be freed before the GL context is
terminated. The solution is to add some explicit `clear()`/`shutdown()`
functions which allow resources to be freed in the correct order.