Add SysMap
SysMap is a (potentially temporary) system which generates a graphical
display of Universe objects (e.g. planets). The Universe objects live in
3D space and are stored with 64-bit integer coordinates. SysMap creates
entities in an ActiveScene corresponding to each Universe object that is
to be drawn on the map. Since we can't render with ints, nor scales up
to 10^19 units, SysMap scales the units such that 1 unit = 1 Megameter
(Mm, 10^6 meters) = 1024*10^6 Universe position units.

Orbital trajectories are drawn by associating each object with a
100-vertex ring which is rendered as a ring. At present the rings are
just circles of fixed radii, but will eventually be proper trajectory
predictions.

Several of the debug classes and main.cpp had to be modified to draw the
map; the modified versions are appended with `2` and the original files
were left alone.