Planet trails
Experimenting with the planets leaving trails. It uses indexed meshes to
reproduce circular buffer behavior in the rendering, although there are
still some issues to be ironed out when the effect is first initialized
and the buffer hasn't yet overwritten its initial data with real data.

Also added the moon, although it isn't very well behaved.