Add lots of solar system moons
Added major moons to mars, jupiter, venus, uranus. Everything runs great
still.

Implemented the ability to choose the starting position of an object by
specifying the polar angle

Added a `TCompAsteroid` component to be used soon in generating a ton of
asteroids, which will read from but not contribute to general system
dynamics.