Optimize non-contributing body dynamics
Asteroids and other bodies which are subject to but do not contribute to
system dynamics can be updated in a very efficient manner by
pre-fetching all the position/mass data from the massive bodies and then
using it to perform a more cache efficient update of all light bodies in
a single pass. Can simulate 5,000 asteroids at about 190fps.