Add new map rendering data class
The new MapRenderData class is responsible for drawing points and lines
in the map screen. It stores a single mesh for the points, and a single
mesh for the paths, and uses indexed rendering to draw them in a single
draw call. Currently only tested with points, paths are WIP