Fix camera focus, one-time map initialization
Moving to a single-entity, single-draw call model ruined the ability to
have the camera focus on scene objects. Rather than spawn a bunch of
`ACompTransform`s and updating them every frame, I just modified the
DebugCameraController to look at the universe's registry instead, and
convert universe coordinates to map coordinates only for the object
being orbited. Works just as well as before.

Separated the initialization of map screen objects into a single
function. Instead of traversing a map every time to check to see if each
entity exists in the graphics output, I just put it in a function that
can be called when necessary. Gives a reasonable perf boost (50fps with
5k asteroids).