Add predicted orbit drawing
For now, trajectories are integrated forwards in time and then revealed
one step at a time. When this data is exhausted, a new future trajectory
is computed and the cycle repeats.

Added a function in TrajNBody which returns true if the last frame
recomputed the future trajectories. This is used to trigger SysMap to
read the current data into its path data (through a vector containing a
single planet's traj data returned from `TrajNBody::get_sat_traj()`)
which is displayed on screen until the next update.