Add new map path manager
Paths are now represented by a pointer-like `MapPath` class which acts
as a handle on a segment of path data. Instead of using a `std::map` to
associate a path to each satellite, paths can now be distributed to any
component so that a body may have multiple paths, or a non-satellite
body may have a path. The MapPaths store an index into a vector of data
allocation structs which contain the indices bounding the relevant path
data in the vertex buffer.

This is a prototype and has a few issues:

* Dynamic systems are freed after the GL context dies, leaving dangling
references to VBOs. The other branch solves this.

* The system ultimately needs a heap-like allocator; right now it just
grows in one direction until it hits the limit.

* The trajectory prediction is currently disabled pending design
decisions about how to make long-term orbit predictions without
simulating 250 years for an orbit of e.g. pluto. May end up calculating
an approximate 2-body orbit, or using conics, to draw quick rings to aid
in long-term intercepts